+++ libunwind/src/EHHeaderParser.hpp
@@ -68,7 +68,7 @@ void EHHeaderParser<A>::decodeEHHdr(A &addressSpace, pint_t ehHdrStart,
   ehHdrInfo.eh_frame_ptr =
       addressSpace.getEncodedP(p, ehHdrEnd, eh_frame_ptr_enc, ehHdrStart);
   ehHdrInfo.fde_count =
-      (size_t)addressSpace.getEncodedP(p, ehHdrEnd, fde_count_enc, ehHdrStart);
+      addressSpace.getEncodedP(p, ehHdrEnd, fde_count_enc, ehHdrStart);
   ehHdrInfo.table = p;
 }
