+++ libunwind/src/config.h
@@ -157,7 +157,7 @@ static inline uintptr_t pcc_address(uintptr_t a)
   if (__builtin_cheri_tag_get((void*)a))
     return a;
   void *pcc = __builtin_cheri_program_counter_get();
-  pcc = __builtin_cheri_offset_set(pcc, (uint64_t)a);
+  pcc = __builtin_cheri_offset_set(pcc, a);
   return (uintptr_t)pcc;
 #else
   return a;
@@ -172,7 +172,7 @@ static inline uintptr_t ddc_address(uintptr_t a)
     return a;
   //fprintf(stderr, "Converting 0x%llx to ddc-relative pointer\n", (unsigned long long)a);
   void *ddc = __builtin_cheri_global_data_get();
-  ddc = __builtin_cheri_offset_set(ddc, (uint64_t)a);
+  ddc = __builtin_cheri_offset_set(ddc, a);
   return (uintptr_t)ddc;
 #else
   return a;
